refactor Recipe.vue -> new function to get recipe from firebase
